+++ dwds/lib/src/debugging/metadata/function.dart
@@ -4,7 +4,7 @@
 
 import 'package:dwds/src/debugging/remote_debugger.dart';
 import 'package:dwds/src/loaders/strategy.dart';
-import 'package:dwds/src/utilities/shared.dart';
+import 'package:dwds/src/utilities/server.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 /// Meta data for a remote Dart function in Chrome.
